requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
    #   wandb
cramjam==2.8.4
    # via fastparquet
dnspython==2.6.1
    # via email-validator
docker-pycreds==0.4.0
    # via wandb
email-validator==2.2.0
    # via fastapi
fastapi==0.115.0
    # via -r requirements.in
fastapi-cli==0.0.5
    # via fastapi
fastparquet==2024.5.0
    # via -r requirements.in
fsspec==2024.9.0
    # via fastparquet
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via wandb
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via fastapi
joblib==1.4.2
    # via scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.1.1
    # via
    #   fastparquet
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
packaging==24.1
    # via fastparquet
pandas==2.2.3
    # via
    #   -r requirements.in
    #   fastparquet
platformdirs==4.3.6
    # via wandb
protobuf==5.28.2
    # via wandb
psutil==6.0.0
    # via wandb
pyarrow==17.0.0
    # via -r requirements.in
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.11
    # via fastapi
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   uvicorn
    #   wandb
requests==2.32.3
    # via wandb
rich==13.8.1
    # via typer
scikit-learn==1.5.2
    # via -r requirements.in
scipy==1.14.1
    # via scikit-learn
sentry-sdk==2.14.0
    # via wandb
setproctitle==1.3.3
    # via wandb
setuptools==75.1.0
    # via wandb
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   docker-pycreds
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.38.6
    # via fastapi
threadpoolctl==3.5.0
    # via scikit-learn
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.31.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.20.0
    # via uvicorn
wandb==0.18.2
    # via -r requirements.in
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
